Move to a mux Router & more containerized code style
